Textbooks have been removed
You can pull without downloading them.
It actually really only took around 15 seconds to pull the pdfs on my
laptop where it took like 3 minutes for it load on google chromes and
then I had to download it from there bc of the site I was using to
download it.