Project 2: Small Tree changes
I am not really that happy with the typing of the tree but I think it
would be difficult to make it better. As I have it right now I need to
do certain runtime checks to ensure that certain node takes are
specfically production because it does not make sense for them to be
terminal. I do not think I can store just the enum type because that is
not wrapped in a mutable reference counter.