Project 2: ??????????????????
When I needed a string token for parsing I realized it did not even
exist... I knew I created it because I made an alphabet test but that
test was not even there and I realized I only had 6 tests but I knew I
did 10. Somehow I lost a couple of commits. This is a bit strange bc I
got full marks on the test cases with string but that's in my favor