Final Project: Work on expression evaluation
Ok so the plan is to have temp values and for boolean expression to
store the output into another temporary "variable". This way I can have
the result of booelan expressions in a temp variable. I can create a
pool of temp values of bool output and digit output. I will only need to
keep track of the index of the pool for the pointers